+++ Packages/com.unity.inputsystem/InputSystem/Plugins/UI/InputSystemUIInputModuleEditor.cs
@@ -23,15 +23,17 @@ namespace UnityEngine.InputSystem.UI.Editor
             return null;
         }
 
-        private static InputActionReference[] GetAllActionsFromAsset(InputActionAsset actions)
+        private static InputActionReference[] GetAllAssetReferencesFromAssetDatabase(InputActionAsset actions)
         {
-            if (actions != null)
-            {
-                var path = AssetDatabase.GetAssetPath(actions);
-                var assets = AssetDatabase.LoadAllAssetsAtPath(path);
-                return assets.Where(asset => asset is InputActionReference).Cast<InputActionReference>().OrderBy(x => x.name).ToArray();
-            }
-            return null;
+            if (actions == null)
+                return null;
+
+            var path = AssetDatabase.GetAssetPath(actions);
+            var assets = AssetDatabase.LoadAllAssetsAtPath(path);
+            return assets.Where(asset => asset is InputActionReference)
+                .Cast<InputActionReference>()
+                .OrderBy(x => x.name)
+                .ToArray();
         }
 
         private static readonly string[] s_ActionNames =
@@ -64,8 +66,14 @@ namespace UnityEngine.InputSystem.UI.Editor
 
         private SerializedProperty[] m_ReferenceProperties;
         private SerializedProperty m_ActionsAsset;
-        private InputActionReference[] m_AvailableActionsInAsset;
+        private InputActionReference[] m_AvailableActionReferencesInAssetDatabase;
         private string[] m_AvailableActionsInAssetNames;
+
+        private string MakeActionReferenceNameUsableInGenericMenu(string name)
+        {
+            // Ugly hack: GenericMenu interprets "/" as a submenu path. But luckily, "/" is not the only slash we have in Unicode.
+            return name.Replace("/", "\uFF0F");
+        }
 
         public void OnEnable()
         {
@@ -75,15 +83,15 @@ namespace UnityEngine.InputSystem.UI.Editor
                 m_ReferenceProperties[i] = serializedObject.FindProperty($"m_{s_ActionNames[i]}Action");
 
             m_ActionsAsset = serializedObject.FindProperty("m_ActionsAsset");
-            m_AvailableActionsInAsset = GetAllActionsFromAsset(m_ActionsAsset.objectReferenceValue as InputActionAsset);
-            // Ugly hack: GenericMenu interprets "/" as a submenu path. But luckily, "/" is not the only slash we have in Unicode.
-            m_AvailableActionsInAssetNames = new[] { "None" }.Concat(m_AvailableActionsInAsset?.Select(x => x.name.Replace("/", "\uFF0F")) ?? new string[0]).ToArray();
+            m_AvailableActionReferencesInAssetDatabase = GetAllAssetReferencesFromAssetDatabase(m_ActionsAsset.objectReferenceValue as InputActionAsset);
+            m_AvailableActionsInAssetNames = new[] { "None" }
+                .Concat(m_AvailableActionReferencesInAssetDatabase?.Select(x => MakeActionReferenceNameUsableInGenericMenu(x.name)) ?? new string[0]).ToArray();
         }
 
         public static void ReassignActions(InputSystemUIInputModule module, InputActionAsset action)
         {
             module.actionsAsset = action;
-            var assets = GetAllActionsFromAsset(action);
+            var assets = GetAllAssetReferencesFromAssetDatabase(action);
             if (assets != null)
             {
                 module.point = GetActionReferenceFromAssets(assets, module.point?.action?.name, "Point", "MousePosition", "Mouse Position");
@@ -122,17 +130,34 @@ namespace UnityEngine.InputSystem.UI.Editor
             }
 
             var numActions = s_ActionNames.Length;
-            for (var i = 0; i < numActions; i++)
+
+            if ((m_AvailableActionReferencesInAssetDatabase != null && m_AvailableActionReferencesInAssetDatabase.Length > 0) || m_ActionsAsset.objectReferenceValue == null)
             {
-                if (m_AvailableActionsInAsset == null)
-                    continue;
+                for (var i = 0; i < numActions; i++)
+                {
+                    var index = Array.IndexOf(m_AvailableActionReferencesInAssetDatabase,
+                        m_ReferenceProperties[i].objectReferenceValue) + 1;
+                    EditorGUI.BeginChangeCheck();
+                    index = EditorGUILayout.Popup(s_ActionNiceNames[i], index, m_AvailableActionsInAssetNames);
 
-                var index = Array.IndexOf(m_AvailableActionsInAsset, m_ReferenceProperties[i].objectReferenceValue) + 1;
-                EditorGUI.BeginChangeCheck();
-                index = EditorGUILayout.Popup(s_ActionNiceNames[i], index, m_AvailableActionsInAssetNames);
+                    if (EditorGUI.EndChangeCheck())
+                        m_ReferenceProperties[i].objectReferenceValue =
+                            index > 0 ? m_AvailableActionReferencesInAssetDatabase[index - 1] : null;
+                }
+            }
+            else
+            {
+                // Somehow we have an asset but no asset references from the database, pull out references manually and show them in read only UI
+                EditorGUILayout.HelpBox("Showing fields as read-only because current action asset seems to be created by a script and assigned programmatically.", MessageType.Info);
+                for (var i = 0; i < numActions; i++)
+                {
+                    var retrievedName = "None";
+                    if (m_ReferenceProperties[i].objectReferenceValue != null &&
+                        (m_ReferenceProperties[i].objectReferenceValue is InputActionReference reference))
+                        retrievedName = MakeActionReferenceNameUsableInGenericMenu(reference.ToNiceName());
 
-                if (EditorGUI.EndChangeCheck())
-                    m_ReferenceProperties[i].objectReferenceValue = index > 0 ? m_AvailableActionsInAsset[index - 1] : null;
+                    EditorGUILayout.Popup(s_ActionNiceNames[i], 0, new[] {retrievedName});
+                }
             }
 
             if (GUI.changed)